moved test for injection to test/test_marin_tokenizer.py
